rename the folder
